Bug fix: Handle unequal-length ranges in `by_word` iterator.
The `by_word` iterator takes the result of `difflib.SequenceMatcher.get_opcodes()` and expands it into an opcode sequence that only modifies one word at a time.   

The code uses `for i1, i2 in zip(range(s1, e1), range(s2, e2))` to expand a “replace” opcode.  But in the case where the ranges have different lengths, `zip()` uses the smaller range, resulting in some words getting dropped from the expansion.

The fix here is to check for that case, and add `’delete’` or `’insert’` operations for the words that weren’t included in the zip.

Fixes #101